+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Enh #41: Add `HasTarget` trait and `target()` method to manage `target` attribute for HTML elements (@terabytesoftw)
 - Enh #42: Add `HasAs` trait and `as()` method to manage `as` attribute for HTML elements (@terabytesoftw)
 - Bug #43: Update copyright year to `2026` in multiple files (@terabytesoftw)
+- Enh #44: Add `HasDisabled` trait and `disabled()` method to manage `disabled` attribute for HTML elements (@terabytesoftw)
 
 ## 0.5.2 January 29, 2026